Rename IteratorsToStandard to IteratorsToLLVM.
