+++ experimental/iterators/lib/Conversion/CMakeLists.txt
@@ -1,1 +1,1 @@
-add_subdirectory(IteratorsToStandard)
+add_subdirectory(IteratorsToLLVM)
